+++ docs/setup/release-notes.md
@@ -67,6 +67,7 @@
 * [<a href='https://github.com/apache/sedona/issues/3288'>GH-3288</a>] - Implement `GeoDataFrame.from_dict` for in-memory dictionaries
 * [<a href='https://github.com/apache/sedona/issues/3344'>GH-3344</a>] - Expose `SpatialIndex.valid_query_predicates`
 * [<a href='https://github.com/apache/sedona/issues/3357'>GH-3357</a>] - Implement `list_layers` for GeoPackage using the native Spark reader
+* [<a href='https://github.com/apache/sedona/issues/3370'>GH-3370</a>] - Add distributed `simplify_coverage` for GeoSeries and GeoDataFrame, using reliable Spark checkpoints
 
 ### Bug Fixes